requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
idna==3.4
joblib==1.3.2
mpmath==1.3.0
mysqlclient==2.2.0
numpy==1.24.4
packaging==23.2
pandas==2.1.2
Pillow==10.0.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
